Add initial WASM wrapper, docs, and Makefile target
Add Emscripten embind bindings and docs to build/run the project as a WebAssembly web app. Introduces src/wasm_wrapper.cpp (wrappers for MD5, Blowfish, and RSA), new documentation files QUICKSTART.md and WASM_BUILD.md, and new Makefile targets (build_wasm, build_emscripten) to produce bin/cryptotool.{html,js,wasm}. Also apply a small type-safety fix in Blowfish::F by adding explicit uint8_t casts when splitting the 32-bit value into bytes.

+++ src/blowfish/blowfish.cpp
@@ -114,10 +114,10 @@ uint32_t Blowfish::F(uint32_t x) {
 
     // Split the 32-bit input into four 8-bit segments
     uint8_t xarr[4] = {
-        x >> 24,
-        x >> 16,
-        x >> 8,
-        x
+        static_cast<uint8_t>(x >> 24),
+        static_cast<uint8_t>(x >> 16),
+        static_cast<uint8_t>(x >> 8),
+        static_cast<uint8_t>(x)
     };
     // Use the segments to index into the S-boxes and combine the results
     uint32_t result;
